Change pad display to grid
I decided to change the pad display from flex-box to grid as it makes
fitting the divs inside the pad container much easier and siginficantly
improves the speed of calculations and therefore the UX.

This also made the pad more responsive so now I no longer need to listen
for window resize events which also improved the overall performance
considerably.

Perhaps I should also change the rest of the layout to grid, but at this
point I'm happy with how it turned out.